+++ core/embed.go
@@ -1,0 +1,6 @@
+package core
+
+import "embed"
+
+//go:embed static
+var staticFS embed.FS