+++ core/router.go
@@ -4,9 +4,7 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"strings"
 
-	assetfs "github.com/elazarl/go-bindata-assetfs"
 	"github.com/gin-contrib/secure"
 	"github.com/gin-contrib/static"
 	"github.com/gin-gonic/gin"
@@ -24,35 +22,6 @@ const (
 // Is this a github repo/org
 var isGithub bool
 
-// binaryFS  holds a filesystem handle
-type binaryFS struct {
-	fs http.FileSystem
-}
-
-// Open will return an http file object that refers to a given file
-func (b *binaryFS) Open(name string) (http.File, error) {
-	return b.fs.Open(name)
-}
-
-// Exists checks if a given file with a given prefix exists and attempts to open it
-func (b *binaryFS) Exists(prefix string, filepath string) bool {
-	if p := strings.TrimPrefix(filepath, prefix); len(p) < len(filepath) {
-		if _, err := b.fs.Open(p); err != nil {
-			return false
-		}
-		return true
-	}
-	return false
-}
-
-// binaryFileSystem returns a binary file system object used by the web frontend
-func binaryFileSystem(root string) *binaryFS {
-	fs := &assetfs.AssetFS{Asset: Asset, AssetDir: AssetDir, AssetInfo: AssetInfo, Prefix: root}
-	return &binaryFS{
-		fs,
-	}
-}
-
 // NewRouter will create an instance of the web frontend, setting the necessary parameters.
 func NewRouter(s *Session) *gin.Engine {
 
@@ -67,7 +36,11 @@ func NewRouter(s *Session) *gin.Engine {
 	}
 
 	router := gin.New()
-	router.Use(static.Serve("/", binaryFileSystem("static")))
+	embedFS, err := static.EmbedFolder(staticFS, "static")
+	if err != nil {
+		s.Out.Fatal("Unable to load embedded static assets: %s", err)
+	}
+	router.Use(static.Serve("/", embedFS))
 	router.Use(secure.New(secure.Config{
 		SSLRedirect:           false,
 		IsDevelopment:         false,